primi fix pr
