+++ frontend/src/components/AddWorkflow/AddWorkflow.test.tsx
@@ -3,7 +3,7 @@ import { vi, describe, test, beforeEach, expect } from "vitest";
 import userEvent from "@testing-library/user-event";
 import { AddWorkflow } from "./AddWorkflow";
 import { useCreateWorkflow } from "../../hooks/useCreateWorkflow";
-import '@testing-library/jest-dom';
+import "@testing-library/jest-dom";
 
 vi.mock("../../hooks/useCreateWorkflow");
