+++ frontend/src/components/WorkflowList/WorkflowList.test.tsx
@@ -1,9 +1,9 @@
 import { screen } from "@testing-library/react";
 import { vi, describe, test, expect, beforeEach } from "vitest";
 import { WorkflowList } from "./WorkflowList";
-import { useAllWorkflow } from "../../hooks/useAllWorkflow";
+import { useAllWorkflow } from "../../hooks/useAllWorkflows";
 import { AuthContextType, authProviderRender, providerPropsInit} from "../../context/MockedAuthProvider"
-import '@testing-library/jest-dom';
+import "@testing-library/jest-dom";
 
 
 // Mock the `useAllWorkflow` hook
