+++ frontend/src/hooks/useAllWorkflows.test.tsx
@@ -1,8 +1,8 @@
 import { expect, test, describe, beforeEach, vi } from "vitest";
-import { render, screen, waitFor, act } from "@testing-library/react";
+import { render, screen, waitFor } from "@testing-library/react";
 import { AuthContextType, authProviderRender, providerPropsInit } from "../context/MockedAuthProvider";
-import { allWorkflow } from "../services/allWorkflowAPI";
-import { useAllWorkflow } from "./useAllWorkflow";
+import { allWorkflow } from "../services/allWorkflowsAPI";
+import { useAllWorkflow } from "./useAllWorkflows";
 import "@testing-library/jest-dom";
 
 vi.mock("../services/allWorkflowAPI", () => ({
