+++ frontend/src/services/constants.ts
@@ -1,0 +1,1 @@
+export const API_BASE_URL = "http://localhost:3000";
